requirements.txt
Django==1.9.2
django-auth-ldap==1.2.7
django-better-filter-widget==0.4
django-bootstrap-form==3.2.1
django-bootstrap3==6.2.2
django-datetime-widget==0.9.3
django-extensions==1.6.1
django-multi-email-field==0.3.1
django-multiupload==0.5
django-multiuploader==0.2.9
psycopg2==2.6.1
python-ldap==2.4.25
python-magic==0.4.11
pytz==2016.1
six==1.10.0
sorl-thumbnail==12.3
wheel==0.24.0
